art study 3: left eye fully rendered (1/2)
oh my GOODNESS was this challenging. it might not look complicated to any who is not an artist but the real challenge here was making sure all the colors, AND proportions of the eyes lined up. you see both eyes have to be identical in size of pupil, shape of pupil, color of iris, and so many other things....oh my goodness...